+++ Glassy/Modules/NewsData.lua
@@ -4,7 +4,7 @@ local Core = unpack(select(2, ...))
 -- luacheck: push ignore 631
 Core.NewsEntries = {
   {
-    name = "1.9.4 (unreleased)",
+    name = "1.9.4 (2026-09-21)",
     items = {[[
 What's new